requirements.txt
anthropic-haystack
haystack-ai
boto3>=1.26.0
botocore>=1.29.0
python-dateutil>=2.8.2
python-dotenv